python/tools/toiftool/pyproject.toml
[tool.poetry]
name = "toiftool"
version = "0.1.0"
description = ""
authors = ["matejcik <ja@matejcik.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
trezor = ">=0.13"
click = ">=7.0,<9.0"
Pillow = "^9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
toiftool = "toiftool:cli"
